Switch amount column type from float to double

+++ tests/amount/test_automapper_amount_typed.py
@@ -51,4 +51,4 @@ def test_auto_mapper_amount_typed(spark_session: SparkSession) -> None:
         result_df.where("member_id == 2").select("age").collect()[0][0]
     ) == approx(67.67)
 
-    assert dict(result_df.dtypes)["age"] == "float"
+    assert dict(result_df.dtypes)["age"] == "double"
